Anhkka123

+++ app/Http/Controllers/ArticleController.php
@@ -11,11 +11,9 @@ use App\Http\Requests\ArticleRequest;
 
 class ArticleController extends Controller
 {
-    //
+    //bla
     public  function table(){
-//        $article = Tintuc::all();
-//        dd($category);
-            $article = Tintuc::orderBy('id','DESC')->get();
+            $article = Tintuc::orderBy('id','DESC')->paginate(5);
         return view('admin.article.TableArticle',['article'=>$article]);
 
     }
@@ -101,7 +99,10 @@ class ArticleController extends Controller
                 $hinh = str_random(4)."_".$name;
             }
             $file->move("upload/News",$hinh);
-            unlink("upload/News/".$article->Hinh);
+            if(file_exists("upload/News/".$article->Hinh)){
+                unlink("upload/News/".$article->Hinh);
+            }
+
             $article->Hinh = $hinh;
         }
